split DigitTwin in two classes
